Introduce core.async in the parser
The parser now spins up a great deal of channels to perform parallel downloads of cards. It currently uses 50 by default, but can be supplied an integer if you want to override it. It takes an option map, meaning that you can call `(all-cards :parallel-limit 1000)` if you want 1000 channels downloading.

This begins the work on issue #9.