Simplified all-cards
The function `all-cards` was turning into a bloated monster. This commit simplifies matters by having `all-cards` return a channel that it then proceeds to put the parsed cards onto. This decouples processing of the cards completely from `all-cards`, which is now only concerned with parsing all cards and putting them on to a channel.

This is a rather different approach from the one mentioned in issue #9, but I think this is much more promising.