Limited the number of parallel downloads
The parser now features a `rate-limit` channel controlling the number of parallel downloads. This prevents the many strange errors that happened when fetching all cards from Gatherer.

NOTE: The limit is currently 100, which is based solely on being a pretty, round number in my base number of choice. It may be worth experimenting with other values.